static cast int.

+++ src/imu/bmi088.cpp
@@ -155,7 +155,8 @@ bool Bmi088::close() { return acc_spi_driver_.close() && gyro_spi_driver_.close(
 int Bmi088::getRaw(std::vector<byte> cmd) { LOG(E, "Bmi088 getRaw not implemented."); }
 
 int8_t Bmi088::readReg(uint8_t reg_addr, uint8_t *reg_data, uint32_t len, void *intf_ptr) {
-  auto res = static_cast<SpiDriver *>(intf_ptr)->xfer({reg_addr}, len, spi_transfer_speed_hz_);
+  auto res = static_cast<SpiDriver *>(intf_ptr)->xfer({reg_addr}, static_cast<int>(len),
+                                                      spi_transfer_speed_hz_);
   std::copy(res.begin(), res.end(), reg_data);
   return res.empty() ? BMI08_E_COM_FAIL : BMI08_OK;
 }
